Implement counter functionality
